requirements.txt
streamlit
pytesseract
pdf2image
pillow
sentence-transformers
requests
numpy

